Extract get interior ranges to separate file
