+++ packages/cursorless-engine/src/processTargets/modifiers/scopeHandlers/CollectionItemScopeHandler/getInteriorRanges.ts
@@ -1,0 +1,30 @@
+import {
+  type Range,
+  type SurroundingPairName,
+  type TextEditor,
+  Position,
+} from "@cursorless/common";
+import type { ScopeHandlerFactory } from "../ScopeHandlerFactory";
+
+export function getInteriorRanges(
+  scopeHandlerFactory: ScopeHandlerFactory,
+  languageId: string,
+  editor: TextEditor,
+  delimiter: SurroundingPairName,
+): Range[] {
+  const scopeHandler = scopeHandlerFactory.create(
+    {
+      type: "surroundingPairInterior",
+      delimiter,
+    },
+    languageId,
+  );
+  return Array.from(
+    scopeHandler.generateScopes(editor, new Position(0, 0), "forward", {
+      containment: undefined,
+      skipAncestorScopes: false,
+      includeDescendantScopes: true,
+    }),
+    (scope) => scope.domain,
+  );
+}
